+++ src/Mvc/Controller/Traits/Abstracts/AbstractRestResponse.php
@@ -68,6 +68,11 @@ trait AbstractRestResponse
     public const string REST_VIEW_COUNT = 'count';
 
     /**
+     * Standard view field containing the request field represented by the data.
+     */
+    public const string REST_VIEW_FIELD = 'field';
+
+    /**
      * Standard view field containing sum aggregate results.
      */
     public const string REST_VIEW_SUM = 'sum';